* 'master' of github.com:grafana/grafana: (4958 commits)
  Chore: Converts HelpModal from angular to react (#19474)
  Fix typo (#19571)
  Chore: Upgrade to Go 1.13 (#19502)
  Explore: Move data source loader into the select (#19465)
  Theme: fixed issue with theme always beeing modified (#19517)
  Release: Fix issue with tag script on osx (#19557)
  Release: Update latest (#19559)
  Docs: Updates about Loki annotations (#19537)
  Theme: follow-up fix for snapshot
  UI: Centers the filter tags in input field (#19546)
  Update README.md (#19515)
  Docs: Updated changelog (#19558)
  Theme: fix theme issue
  Provisioning: Handle empty nested keys on YAML provisioning datasources (#19547)
  Docs: updates to what's new in 6.4 (#19539)
  Loki: remove live option for logs panel (#19533)
  Chore: Updates to 6.4.0 stable (#19528)
  CloudWatch: Add ap-east-1 to hard-coded region lists (#19523)
  ChangeLog: Release 6.4.0 Stable (#19526)
  Docs: Add notice about plugins that need updating (#19519)
  ...

+++ pkg/tsdb/stackdriver/types.go
@@ -26,7 +26,7 @@ type StackdriverBucketOptions struct {
 		Scale            float64 `json:"scale"`
 	} `json:"exponentialBuckets"`
 	ExplicitBuckets *struct {
-		Bounds []int64 `json:"bounds"`
+		Bounds []float64 `json:"bounds"`
 	} `json:"explicitBuckets"`
 }
